eld_app/requirements.txt
django==4.2
djangorestframework==3.14
geopy==2.3
requests==2.28
polyline==2.0
gunicorn==20.1